api/requirements.txt
fastapi
uvicorn
python-multipart
langchain-google-genai
pypdf
python-dotenv
langchain-core
langchain-community
